# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"CodeAtlas API","description":"API for the AI-Powered CodeAtlas Platform","version":"1.0.0"},"paths":{"/api/v1/auth/signup":{"post":{"tags":["auth"],"summary":"Create User","description":"Create new user.","operationId":"create_user_api_v1_auth_signup_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/login/access-token":{"post":{"tags":["auth"],"summary":"Login Access Token","description":"OAuth2 compatible token login, get an access token for future requests","operationId":"login_access_token_api_v1_auth_login_access_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_access_token_api_v1_auth_login_access_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/users/me":{"get":{"tags":["users"],"summary":"Read User Me","description":"Get current user profile.","operationId":"read_user_me_api_v1_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Read User Me Api V1 Users Me Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/roadmap/latest":{"get":{"tags":["roadmap"],"summary":"Get Latest Roadmap","operationId":"get_latest_roadmap_api_v1_roadmap_latest_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Latest Roadmap Api V1 Roadmap Latest Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/roadmap/":{"get":{"tags":["roadmap"],"summary":"Get All Roadmaps","description":"List all roadmaps for the current user.","operationId":"get_all_roadmaps_api_v1_roadmap__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get All Roadmaps Api V1 Roadmap  Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/roadmap/{roadmap_id}":{"get":{"tags":["roadmap"],"summary":"Get Roadmap By Id","description":"Get a specific roadmap by ID.","operationId":"get_roadmap_by_id_api_v1_roadmap__roadmap_id__get","parameters":[{"name":"roadmap_id","in":"path","required":true,"schema":{"type":"string","title":"Roadmap Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Roadmap By Id Api V1 Roadmap  Roadmap Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/roadmap/{roadmap_id}/bookmark":{"put":{"tags":["roadmap"],"summary":"Toggle Bookmark","description":"Toggle bookmark status.","operationId":"toggle_bookmark_api_v1_roadmap__roadmap_id__bookmark_put","parameters":[{"name":"roadmap_id","in":"path","required":true,"schema":{"type":"string","title":"Roadmap Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Toggle Bookmark Api V1 Roadmap  Roadmap Id  Bookmark Put"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/roadmap/generate":{"post":{"tags":["roadmap"],"summary":"Generate Roadmap","operationId":"generate_roadmap_api_v1_roadmap_generate_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"x-groq-api-key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Groq-Api-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Body"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Generate Roadmap Api V1 Roadmap Generate Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/roadmap/cleanup":{"delete":{"tags":["roadmap"],"summary":"Cleanup Non Bookmarked","description":"Delete all roadmaps for user that are NOT bookmarked.\nCalled on SignOut.","operationId":"cleanup_non_bookmarked_api_v1_roadmap_cleanup_delete","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Cleanup Non Bookmarked Api V1 Roadmap Cleanup Delete"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/dashboard/stats":{"get":{"tags":["dashboard"],"summary":"Get User Stats","description":"Get real-time user statistics with detailed breakdown.","operationId":"get_user_stats_api_v1_dashboard_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get User Stats Api V1 Dashboard Stats Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/dashboard/solve/{problem_id}":{"post":{"tags":["dashboard"],"summary":"Toggle Problem","description":"Toggle a problem as solved/unsolved.\nUpdates streak and solved count.","operationId":"toggle_problem_api_v1_dashboard_solve__problem_id__post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"problem_id","in":"path","required":true,"schema":{"type":"string","title":"Problem Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Toggle Problem Api V1 Dashboard Solve  Problem Id  Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/dashboard/progress":{"post":{"tags":["dashboard"],"summary":"Update Progress","description":"Bulk update solved problems (Manual Save).\nReplaces the entire list of solved problems and recalculates streaks.","operationId":"update_progress_api_v1_dashboard_progress_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Item"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Update Progress Api V1 Dashboard Progress Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/interview/resume/upload":{"post":{"tags":["interview"],"summary":"Upload Resume","description":"Upload and parse a PDF resume. Returns extracted text and structured context.","operationId":"upload_resume_api_v1_interview_resume_upload_post","parameters":[{"name":"x-groq-api-key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Groq-Api-Key"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_resume_api_v1_interview_resume_upload_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Upload Resume Api V1 Interview Resume Upload Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/interview/start":{"post":{"tags":["interview"],"summary":"Start Interview","description":"Start a new AI Interview Session.\nInputs: role, company, type, resume_text (optional), resume_context (optional)","operationId":"start_interview_api_v1_interview_start_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"x-groq-api-key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Groq-Api-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Body"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Start Interview Api V1 Interview Start Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/interview/{session_id}/reply":{"post":{"tags":["interview"],"summary":"Reply To Interview","description":"Process user answer and get next question.","operationId":"reply_to_interview_api_v1_interview__session_id__reply_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","title":"Session Id"}},{"name":"x-groq-api-key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Groq-Api-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Body"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Reply To Interview Api V1 Interview  Session Id  Reply Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/interview/history":{"get":{"tags":["interview"],"summary":"Get Interview History","operationId":"get_interview_history_api_v1_interview_history_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Interview History Api V1 Interview History Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/analytics/kpis":{"get":{"tags":["analytics"],"summary":"Get Kpis","description":"Get high-level KPIs for the analytics dashboard","operationId":"get_kpis_api_v1_analytics_kpis_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Kpis Api V1 Analytics Kpis Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/analytics/skills":{"get":{"tags":["analytics"],"summary":"Get Skill Distribution","description":"Get skill breakdown (e.g. by Track)","operationId":"get_skill_distribution_api_v1_analytics_skills_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Skill Distribution Api V1 Analytics Skills Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/analytics/interviews":{"get":{"tags":["analytics"],"summary":"Get Interview Trends","description":"Get recent interview performance trend","operationId":"get_interview_trends_api_v1_analytics_interviews_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Interview Trends Api V1 Analytics Interviews Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/analytics/feedback":{"get":{"tags":["analytics"],"summary":"Get Recent Feedback","description":"Get aggregated feedback from recent sessions.","operationId":"get_recent_feedback_api_v1_analytics_feedback_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Recent Feedback Api V1 Analytics Feedback Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/vault/items":{"get":{"tags":["vault"],"summary":"Get Vault Items","description":"Get all proof items for the current user.","operationId":"get_vault_items_api_v1_vault_items_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Vault Items Api V1 Vault Items Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["vault"],"summary":"Create Vault Item","description":"Add a new proof item to the vault.","operationId":"create_vault_item_api_v1_vault_items_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Item"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Create Vault Item Api V1 Vault Items Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/vault/items/{item_id}":{"delete":{"tags":["vault"],"summary":"Delete Vault Item","description":"Remove a proof item.","operationId":"delete_vault_item_api_v1_vault_items__item_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"item_id","in":"path","required":true,"schema":{"type":"string","title":"Item Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Delete Vault Item Api V1 Vault Items  Item Id  Delete"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/project/visualize":{"post":{"tags":["project"],"summary":"Project Visualize","description":"Step 1: Get Mermaid Graph Data.","operationId":"project_visualize_api_v1_project_visualize_post","parameters":[{"name":"x-groq-api-key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Groq-Api-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Body"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Project Visualize Api V1 Project Visualize Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/project/docs":{"post":{"tags":["project"],"summary":"Project Docs","description":"Step 2: Generate README & API Specs.","operationId":"project_docs_api_v1_project_docs_post","parameters":[{"name":"x-groq-api-key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Groq-Api-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Body"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Project Docs Api V1 Project Docs Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_login_access_token_api_v1_auth_login_access_token_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_access_token_api_v1_auth_login_access_token_post"},"Body_upload_resume_api_v1_interview_resume_upload_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_resume_api_v1_interview_resume_upload_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["access_token","token_type"],"title":"Token"},"User":{"properties":{"email":{"anyOf":[{"type":"string","format":"email"},{"type":"null"}],"title":"Email"},"is_active":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Active","default":true},"full_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Name"},"target_role":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Target Role"},"id":{"type":"string","title":"Id"}},"type":"object","required":["id"],"title":"User"},"UserCreate":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"is_active":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Active","default":true},"full_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Name"},"target_role":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Target Role"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserCreate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/v1/auth/login/access-token"}}}}}}